requirements.txt
python-telegram-bot==22.3
openai==2.7.1
reportlab==4.2.5
